Protected routes, added greeting, handled errors
The index.js routes were edited with the new MW for authenitcation. Sessions was used to prevent users from being logged out on refresh. The notes functionality was started using .lean() to link the template to server and error handling templates were made for 404 and 500 errors. Lastly, a loop was created to loop through all notes and output them to a table.